fix: rename pkg prepare; move dir ut to tatris/test/
