+++ internal/query/handler/query_handler_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"github.com/tatris-io/tatris/internal/common/consts"
-	prepare2 "github.com/tatris-io/tatris/internal/ut/prepare"
+	"github.com/tatris-io/tatris/test/ut/prepare"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -20,7 +20,7 @@ import (
 func TestQueryHandler(t *testing.T) {
 
 	// prepare
-	index, _, err := prepare2.CreateIndexAndDocs(time.Now().Format(consts.VersionTimeFmt))
+	index, _, err := prepare.CreateIndexAndDocs(time.Now().Format(consts.VersionTimeFmt))
 	if err != nil {
 		t.Fatalf("prepare index and docs fail: %s", err.Error())
 	}
